Update report with some explanation of how the app should work.
Implemented canDraw in draw part as argument of constructor,
this way the player can only exit the panel if it is not the leader.
Defined the first leader as the first member of the channel.
Defined isLeader function and used to define canDraw of the drawPanel,
thus letting only leader with drawing capabilities. Used isLeader to
define the command 'start' - only accessible to the leader.
Used isLeader to give special treatment to submitted answers.